pyproject.toml
[project]
name = "hello-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "google-generativeai",
    "PyPDF2",
    "requests",
    "python-dotenv",
    "youtube-transcript-api",
]
